reorganize imports and move functions
