+++ modules/adnl_over_tcp.py
@@ -1,14 +1,73 @@
 #!/usr/bin/env python3
 # -*- coding: utf_8 -*-
 
+import json
 import time
 import types
+import pytoniq
 from mypylib import Dict
-
-from utils import get_lite_balancer
-
+from utils import get_module_by_name
 
 COMMENT_OP = bytes.fromhex("00000000")
+
+
+def get_lite_balancer(local):
+	main_module = get_module_by_name(local, "main")
+	with open(main_module.global_config_path, "r") as f:
+		config = json.load(f)
+	client = pytoniq.LiteBalancer.from_config(config, trust_level=2)
+	return client
+#end define
+
+def normalize_msg_hash(message):
+	if not message.is_external:
+		return message.serialize().hash
+	cell = (
+		pytoniq.begin_cell()
+		.store_uint(2, 2)
+		.store_address(None)
+		.store_address(message.info.dest)
+		.store_coins(0)
+		.store_bool(False)
+		.store_bool(True)
+		.store_ref(message.body)
+		.end_cell()
+	)
+	return cell.hash.hex()
+#end define
+
+async def create_wallet_transfer_payload(wallet, destination, amount, body=None):
+	await wallet.obj.update()
+	if wallet.obj.account.is_uninitialized():
+		seqno = 0
+		state_init = wallet.obj.state_init
+	else:
+		seqno = await wallet.obj.get_seqno()
+		state_init = None
+	if isinstance(destination, str):
+		destination = pytoniq.Address(destination)
+	msg = wallet.obj.create_wallet_internal_message(
+		destination=pytoniq.Address(destination),
+		value=int(amount * 1e9),
+		body=body
+	)
+	body = wallet.obj.raw_create_transfer_msg(
+		private_key=wallet.obj.private_key,
+		seqno=seqno,
+		wallet_id=wallet.obj.wallet_id,
+		messages=[msg]
+	)
+	message = wallet.obj.create_external_msg(
+		src=None,
+		dest=wallet.obj.address,
+		import_fee=0,
+		body=body,
+		state_init=state_init,
+	)
+	msg_boc = message.serialize().to_boc()
+	msg_hash = normalize_msg_hash(message)
+	return msg_boc, msg_hash
+#end define
 
 async def get_account(local, *args, **kwargs):
 	client = get_lite_balancer(local)
